+++ sdk_test/type2_text_test.cpp
@@ -85,9 +85,9 @@ TEST_F(Type2TextTest, Save_Then_Load) {
   ASSERT_TRUE(CreateMsgTextFile());
 
   auto m1 = save_message("Hello World");
-  ASSERT_EQ(1, m1->stored_as);
+  ASSERT_EQ(1u, m1->stored_as);
   auto m2 = save_message("Hello World2");
-  ASSERT_EQ(2, m2->stored_as);
+  ASSERT_EQ(2u, m2->stored_as);
 
   auto out = readfile(m1.value());
   ASSERT_EQ("Hello World", *out);
@@ -100,14 +100,14 @@ TEST_F(Type2TextTest, TwoBlocks) {
   ASSERT_TRUE(CreateMsgTextFile());
 
   auto m1 = save_message("Hello World");
-  ASSERT_EQ(1, m1->stored_as);
+  ASSERT_EQ(1u, m1->stored_as);
 
   const std::string two_blocks(513, 'x');
   auto m2 = save_message(two_blocks);
-  ASSERT_EQ(2, m2->stored_as);
+  ASSERT_EQ(2u, m2->stored_as);
 
   auto m4 = save_message("Hello World");
-  ASSERT_EQ(4, m4->stored_as);
+  ASSERT_EQ(4u, m4->stored_as);
 
   auto out = readfile(m2.value());
   ASSERT_EQ(two_blocks, *out);
@@ -117,13 +117,13 @@ TEST_F(Type2TextTest, Reuse_Block_After_Delete) {
   ASSERT_TRUE(CreateMsgTextFile());
 
   auto m1 = save_message("Hello World");
-  ASSERT_EQ(1, m1->stored_as);
+  ASSERT_EQ(1u, m1->stored_as);
   auto m2 = save_message("Hello World2");
-  ASSERT_EQ(2, m2->stored_as);
+  ASSERT_EQ(2u, m2->stored_as);
 
   ASSERT_TRUE(t_->remove_link(m1.value()));
   auto m3 = save_message("Hello World3");
-  ASSERT_EQ(1, m3->stored_as);
+  ASSERT_EQ(1u, m3->stored_as);
 
 }
 
@@ -134,18 +134,18 @@ TEST_F(Type2TextTest, Move_To_Next_Gat_Section) {
   // Fill up first section.  First section will have only 63 free blocks
   for (auto i = 0; i < 31; i++) {
     auto m = save_message(msg32k);
-    ASSERT_EQ((i * 64) + 1, m->stored_as);
+    ASSERT_EQ(static_cast<uint32_t>(i * 64) + 1, m->stored_as);
   }
 
   // Need 64 blocks, so will start in 2nd section.
   auto m2 = save_message(msg32k);
-  ASSERT_EQ(1, m2->stored_as / 2048);
-  ASSERT_EQ(1, m2->stored_as % 2048);
+  ASSERT_EQ(1u, m2->stored_as / 2048);
+  ASSERT_EQ(1u, m2->stored_as % 2048);
 
   // Need 1 blocks, so will start in 1nd section again.
   auto m3 = save_message("Hello World3");
-  ASSERT_EQ(0, m3->stored_as / 2048);
-  ASSERT_EQ(1985, m3->stored_as % 2048);
+  ASSERT_EQ(0u, m3->stored_as / 2048);
+  ASSERT_EQ(1985u, m3->stored_as % 2048);
 }
